requirements.txt
PyQt6>=6.0.0
langchain>=0.1.0
langchain-community>=0.0.20
openai>=1.0.0
sumy>=0.11.0
sentence-transformers==2.5.*
faiss-cpu
pypdf
python-docx
python-pptx
watchdog
rapidfuzz
chardet
